Extract libzfs error
Extract the libzfs error so it's usable
without linking against libzfs.
